refactor : Move files new directory structure
