+++ app/src/main/java/campus/tech/kakao/map/data/Place.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.model
+package campus.tech.kakao.map.data
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
