+++ app/src/main/java/campus/tech/kakao/map/data/PlaceRepository.kt
@@ -1,6 +1,7 @@
-package campus.tech.kakao.map.repository
+package campus.tech.kakao.map.data
 
-import campus.tech.kakao.map.model.Place
+import campus.tech.kakao.map.api.KakaoApiDataSource
+import campus.tech.kakao.map.data.Place
 import javax.inject.Inject
 
 class PlaceRepository @Inject constructor(private val kakaoApiDataSource : KakaoApiDataSource){
