+++ app/src/main/java/campus/tech/kakao/map/data/SavedPlace.kt
@@ -1,8 +1,8 @@
-package campus.tech.kakao.map.model
+package campus.tech.kakao.map.data
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import campus.tech.kakao.map.db.PlaceContract
+import campus.tech.kakao.map.utilities.PlaceContract
 
 @Entity(tableName = PlaceContract.SavedPlaceEntry.TABLE_NAME)
 data class SavedPlace(
