+++ app/src/main/java/campus/tech/kakao/map/data/SavedPlaceDao.kt
@@ -1,10 +1,8 @@
-package campus.tech.kakao.map.model
+package campus.tech.kakao.map.data
 
 import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import campus.tech.kakao.map.db.PlaceContract
+import campus.tech.kakao.map.utilities.PlaceContract
 
 @Dao
 interface SavedPlaceDao {
