+++ app/src/main/java/campus/tech/kakao/map/di/NetworkModule.kt
@@ -1,11 +1,9 @@
-package campus.tech.kakao.map.model
+package campus.tech.kakao.map.di
 
-import android.content.Context
-import campus.tech.kakao.map.repository.KakaoApiDataSource
+import campus.tech.kakao.map.api.KakaoApiDataSource
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
