+++ app/src/main/java/campus/tech/kakao/map/utilities/PlaceContract.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.db
+package campus.tech.kakao.map.utilities
 
 import android.provider.BaseColumns
