+++ app/src/main/java/campus/tech/kakao/map/view/ClickListener.kt
@@ -1,7 +1,7 @@
 package campus.tech.kakao.map.view
 
-import campus.tech.kakao.map.model.Place
-import campus.tech.kakao.map.model.SavedPlace
+import campus.tech.kakao.map.data.Place
+import campus.tech.kakao.map.data.SavedPlace
 
 interface OnClickPlaceListener {
     fun savePlace(place: Place)
