+++ app/src/main/java/campus/tech/kakao/map/view/PlaceViewAdapter.kt
@@ -1,20 +1,15 @@
-package campus.tech.kakao.map.adapter
+package campus.tech.kakao.map.view
 
-import android.util.Log
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.annotation.VisibleForTesting
 import androidx.databinding.DataBindingUtil
-import androidx.lifecycle.LiveData
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import campus.tech.kakao.map.view.OnClickPlaceListener
 import campus.tech.kakao.map.R
 import campus.tech.kakao.map.databinding.PlaceItemBinding
-import campus.tech.kakao.map.model.Place
+import campus.tech.kakao.map.data.Place
 
 class PlaceViewAdapter(
     val listener: OnClickPlaceListener
