+++ app/src/main/res/layout/bottom_sheet.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="place"
-            type="campus.tech.kakao.map.model.Place"/>
+            type="campus.tech.kakao.map.data.Place"/>
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
