+++ app/src/main/res/layout/saved_place_item.xml
@@ -7,7 +7,7 @@
     <data>
         <variable
             name="savedPlace"
-            type="campus.tech.kakao.map.model.SavedPlace"/>
+            type="campus.tech.kakao.map.data.SavedPlace"/>
 
         <variable
             name="listener"
